+++ policy_set_integration_test.go
@@ -27,9 +27,9 @@ func TestPolicySetsList(t *testing.T) {
 	workspace, workspaceCleanup := createWorkspace(t, client, orgTest)
 	defer workspaceCleanup()
 
-	psTest1, psTestCleanup1 := createPolicySet(t, client, orgTest, nil, []*Workspace{workspace}, "")
+	psTest1, psTestCleanup1 := createPolicySet(t, client, orgTest, nil, []*Workspace{workspace}, nil, "")
 	defer psTestCleanup1()
-	psTest2, psTestCleanup2 := createPolicySet(t, client, orgTest, nil, []*Workspace{workspace}, "")
+	psTest2, psTestCleanup2 := createPolicySet(t, client, orgTest, nil, []*Workspace{workspace}, nil, "")
 	defer psTestCleanup2()
 
 	t.Run("without list options", func(t *testing.T) {
@@ -373,7 +373,7 @@ func TestPolicySetsRead(t *testing.T) {
 
 	upgradeOrganizationSubscription(t, client, orgTest)
 
-	psTest, psTestCleanup := createPolicySet(t, client, orgTest, nil, nil, "")
+	psTest, psTestCleanup := createPolicySet(t, client, orgTest, nil, nil, nil, "")
 	defer psTestCleanup()
 
 	t.Run("with a valid ID", func(t *testing.T) {
@@ -443,7 +443,7 @@ func TestPolicySetsUpdate(t *testing.T) {
 
 	upgradeOrganizationSubscription(t, client, orgTest)
 
-	psTest, psTestCleanup := createPolicySet(t, client, orgTest, nil, nil, "")
+	psTest, psTestCleanup := createPolicySet(t, client, orgTest, nil, nil, nil, "")
 	defer psTestCleanup()
 
 	t.Run("with valid attributes", func(t *testing.T) {
@@ -491,7 +491,7 @@ func TestPolicySetsAddPolicies(t *testing.T) {
 	defer pTestCleanup1()
 	pTest2, pTestCleanup2 := createPolicy(t, client, orgTest)
 	defer pTestCleanup2()
-	psTest, psTestCleanup := createPolicySet(t, client, orgTest, nil, nil, "")
+	psTest, psTestCleanup := createPolicySet(t, client, orgTest, nil, nil, nil, "")
 	defer psTestCleanup()
 
 	t.Run("with policies provided", func(t *testing.T) {
@@ -546,7 +546,7 @@ func TestPolicySetsRemovePolicies(t *testing.T) {
 	defer pTestCleanup1()
 	pTest2, pTestCleanup2 := createPolicy(t, client, orgTest)
 	defer pTestCleanup2()
-	psTest, psTestCleanup := createPolicySet(t, client, orgTest, []*Policy{pTest1, pTest2}, nil, "")
+	psTest, psTestCleanup := createPolicySet(t, client, orgTest, nil, nil, nil, "")
 	defer psTestCleanup()
 
 	t.Run("with policies provided", func(t *testing.T) {
@@ -595,7 +595,7 @@ func TestPolicySetsAddWorkspaces(t *testing.T) {
 	defer wTestCleanup1()
 	wTest2, wTestCleanup2 := createWorkspace(t, client, orgTest)
 	defer wTestCleanup2()
-	psTest, psTestCleanup := createPolicySet(t, client, orgTest, nil, nil, "")
+	psTest, psTestCleanup := createPolicySet(t, client, orgTest, nil, nil, nil, "")
 	defer psTestCleanup()
 
 	t.Run("with workspaces provided", func(t *testing.T) {
@@ -664,7 +664,7 @@ func TestPolicySetsRemoveWorkspaces(t *testing.T) {
 	defer wTestCleanup1()
 	wTest2, wTestCleanup2 := createWorkspace(t, client, orgTest)
 	defer wTestCleanup2()
-	psTest, psTestCleanup := createPolicySet(t, client, orgTest, nil, []*Workspace{wTest1, wTest2}, "")
+	psTest, psTestCleanup := createPolicySet(t, client, orgTest, nil, []*Workspace{wTest1, wTest2}, nil, "")
 	defer psTestCleanup()
 
 	t.Run("with workspaces provided", func(t *testing.T) {
@@ -714,6 +714,140 @@ func TestPolicySetsRemoveWorkspaces(t *testing.T) {
 	})
 }
 
+func TestPolicySetsAddProjects(t *testing.T) {
+	skipUnlessBeta(t)
+	client := testClient(t)
+	ctx := context.Background()
+
+	orgTest, orgTestCleanup := createOrganization(t, client)
+	defer orgTestCleanup()
+
+	upgradeOrganizationSubscription(t, client, orgTest)
+
+	pTest1, pTestCleanup1 := createProject(t, client, orgTest)
+	defer pTestCleanup1()
+	pTest2, pTestCleanup2 := createProject(t, client, orgTest)
+	defer pTestCleanup2()
+	psTest, psTestCleanup := createPolicySet(t, client, orgTest, nil, nil, nil, "")
+	defer psTestCleanup()
+
+	t.Run("with projects provided", func(t *testing.T) {
+		err := client.PolicySets.AddProjects(
+			ctx,
+			psTest.ID,
+			PolicySetAddProjectsOptions{
+				Projects: []*Project{pTest1, pTest2},
+			},
+		)
+		require.NoError(t, err)
+
+		ps, err := client.PolicySets.Read(ctx, psTest.ID)
+		require.NoError(t, err)
+		assert.Equal(t, 2, ps.ProjectCount)
+
+		ids := []string{}
+		for _, ws := range ps.Projects {
+			ids = append(ids, ws.ID)
+		}
+
+		assert.Contains(t, ids, pTest1.ID)
+		assert.Contains(t, ids, pTest2.ID)
+	})
+
+	t.Run("without projects provided", func(t *testing.T) {
+		err := client.PolicySets.AddProjects(
+			ctx,
+			psTest.ID,
+			PolicySetAddProjectsOptions{},
+		)
+		assert.Equal(t, err, ErrRequiredProject)
+	})
+
+	t.Run("with empty projects slice", func(t *testing.T) {
+		err := client.PolicySets.AddProjects(
+			ctx,
+			psTest.ID,
+			PolicySetAddProjectsOptions{Projects: []*Project{}},
+		)
+		assert.Equal(t, err, ErrProjectMinLimit)
+	})
+
+	t.Run("without a valid ID", func(t *testing.T) {
+		err := client.PolicySets.AddProjects(
+			ctx,
+			badIdentifier,
+			PolicySetAddProjectsOptions{
+				Projects: []*Project{pTest1, pTest2},
+			},
+		)
+		assert.Equal(t, err, ErrInvalidPolicySetID)
+	})
+}
+
+func TestPolicySetsRemoveProjects(t *testing.T) {
+	skipUnlessBeta(t)
+	client := testClient(t)
+	ctx := context.Background()
+
+	orgTest, orgTestCleanup := createOrganization(t, client)
+	defer orgTestCleanup()
+
+	upgradeOrganizationSubscription(t, client, orgTest)
+
+	pTest1, pTestCleanup1 := createProject(t, client, orgTest)
+	defer pTestCleanup1()
+	pTest2, pTestCleanup2 := createProject(t, client, orgTest)
+	defer pTestCleanup2()
+	psTest, psTestCleanup := createPolicySet(t, client, orgTest, nil, nil, []*Project{pTest1, pTest2}, "")
+	defer psTestCleanup()
+
+	t.Run("with projects provided", func(t *testing.T) {
+		err := client.PolicySets.RemoveProjects(
+			ctx,
+			psTest.ID,
+			PolicySetRemoveProjectsOptions{
+				Projects: []*Project{pTest1, pTest2},
+			},
+		)
+		require.NoError(t, err)
+
+		ps, err := client.PolicySets.Read(ctx, psTest.ID)
+		require.NoError(t, err)
+
+		assert.Equal(t, 0, ps.ProjectCount)
+		assert.Empty(t, ps.Projects)
+	})
+
+	t.Run("without projects provided", func(t *testing.T) {
+		err := client.PolicySets.RemoveProjects(
+			ctx,
+			psTest.ID,
+			PolicySetRemoveProjectsOptions{},
+		)
+		assert.Equal(t, err, ErrRequiredProject)
+	})
+
+	t.Run("with empty projects slice", func(t *testing.T) {
+		err := client.PolicySets.RemoveProjects(
+			ctx,
+			psTest.ID,
+			PolicySetRemoveProjectsOptions{Projects: []*Project{}},
+		)
+		assert.Equal(t, err, ErrProjectMinLimit)
+	})
+
+	t.Run("without a valid ID", func(t *testing.T) {
+		err := client.PolicySets.RemoveProjects(
+			ctx,
+			badIdentifier,
+			PolicySetRemoveProjectsOptions{
+				Projects: []*Project{pTest1, pTest2},
+			},
+		)
+		assert.Equal(t, err, ErrInvalidPolicySetID)
+	})
+}
+
 func TestPolicySetsDelete(t *testing.T) {
 	client := testClient(t)
 	ctx := context.Background()
@@ -723,7 +857,7 @@ func TestPolicySetsDelete(t *testing.T) {
 
 	upgradeOrganizationSubscription(t, client, orgTest)
 
-	psTest, _ := createPolicySet(t, client, orgTest, nil, nil, "")
+	psTest, _ := createPolicySet(t, client, orgTest, nil, nil, nil, "")
 
 	t.Run("with valid options", func(t *testing.T) {
 		err := client.PolicySets.Delete(ctx, psTest.ID)